requirements-dev.txt
pytest>=8
ruff>=0.8
jsonschema>=4.23
PyYAML>=6
